大幅リファクタリング・途中

+++ HelloGC/src/classes/character/GameCharacterBasePoint.java
@@ -3,49 +3,49 @@ package classes.character;
 public enum GameCharacterBasePoint {
 	CENTER {
 		@Override
-		public int getX(int x, int width) {
+		public float getX(float x, int width) {
 			return x;
 		}
 
 		@Override
-		public int getY(int y, int height) {
+		public float getY(float y, int height) {
 			return y;
 		}
 	},
 	LEFTTOP {
 		@Override
-		public int getX(int x, int width) {
+		public float getX(float x, int width) {
 			return x + (width / 2);
 		}
 
 		@Override
-		public int getY(int y, int height) {
+		public float getY(float y, int height) {
 			return y - (height / 2);
 		}
 	},
 	LEFTBOTTOM {
 		@Override
-		public int getX(int x, int width) {
+		public float getX(float x, int width) {
 			return x + (width / 2);
 		}
 
 		@Override
-		public int getY(int y, int height) {
+		public float getY(float y, int height) {
 			return y + (height / 2);
 		}
 	},
 	RIGHTTOP {
 		@Override
-		public int getX(int x, int width) {
+		public float getX(float x, int width) {
 			return x - (width / 2);
 		}
 
 		@Override
-		public int getY(int y, int height) {
+		public float getY(float y, int height) {
 			return y - (height / 2);
 		}
 	};
-	abstract public int getX(int x, int width);
+	abstract public float getX(float x, int width);
 
-	abstract public int getY(int y, int height);
+	abstract public float getY(float y, int height);
 }
